backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pymongo
pydantic
python-dotenv
python-multipart
motor